pyproject.toml
[tool.poetry]
name = "tagsense"
version = "0.1.0"
description = ""
authors = [
    "Steven Broaddus",
    "Guilherme Oliveira",
    "Spencer Hurt",
    "Jerrin Wofford",
    "Scott Caley",
    "Grant McGeehen",
]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
requests = "^2.32.3"
pyqt6 = "^6.4.2"
markdown = "^3.7"
pillow = "^11.1.0"
sqlparse = "^0.5.3"
langchain-openai = "^0.3.7"
langchain = "^0.3.19"
dotenv = "^0.9.9"
python-dotenv = "^1.0.1"
jsonschema = "^4.23.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
